cattle-streaming/requirements.txt
paho-mqtt==2.1.0
python-dotenv==1.0.1
pymongo==4.10.1
